+++ lib/active_record_filter/component.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module ActiveFilter
+module ActiveRecordFilter
   # Base component that other filter components should extend.
   # A component must implement `filter` that returns an ActiveRecord relation.
   #