+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.0.1
 
- - Introduced `ActiveFilter::Base` and `ActiveFilter::Component`
+ - Introduced `ActiveRecordFilter::Base` and `ActiveRecordFilter::Component`